+++ NEWS.md
@@ -1,5 +1,8 @@
 # vistime (development version)
 
+## New features
+- `vistime()` passes `source` and `customdata` to plot for interactive `Plotly` click or hover events in Shiny. Thanks @jonocarroll.
+
 ## Bugfixes
 - Now asserts that `col.color` and `col.fontcolor` are string type, i.e. `vistime(dat, col.color = color_column)` fails and `vistime(dat, col.color = "color_column")` succeeds. This was implemented for the other columns already.